feat: allow user data to be passed to event handlers
When an event is triggered on a context, that trigger point can now
pass some custom information through to all event handlers which,
for example, provides additional context that is particular to that
trigger point.

At each main phase in the request lifecycle, a "casaRequestPhase"
attribute is sent to event handlers whenever "putContext()" is called,
which allows those handlers to decide the most appropriate action
to take.

Pre-Commit-Config: 6005b639db5362cb1f9354e91d873f310ba9174fb5b4a34588229dd146a5483a

+++ examples/barebones/definitions/events.js
@@ -1,5 +1,7 @@
 // Some examples of how events can be used to manipulate the JourneyContext
 
+import { constants } from '@dwp/govuk-casa';
+
 export default (plan) => [
   // This "waypoint-change" event is triggered whenever data/validation
   // changes on a waypoint, just prior to it being committed to the session.
@@ -30,8 +32,23 @@ export default (plan) => [
   // pinpoint the waypoints you want to purge.
   {
     event: 'context-change',
-    handler: ({ journeyContext }) => {
+    handler: ({ journeyContext, userInfo }) => {
+      // Ignore when the context is updated at the "gather" stage, because it
+      // will always be invalidated at that point. We instead need to wait until
+      // the data has been validated before we test traversals here.
+      if (userInfo?.casaRequestPhase === constants.REQUEST_PHASE_GATHER) {
+        console.log('Skipping purge at the "gather" phase');
+        return;
+      }
+
+      // If the last traversed waypoint has errors, ignore the event and allow
+      // the user to correct errors before we purge.
       const traversed = plan.traverse(journeyContext);
+      if (!traversed.length || journeyContext.getValidationErrorsForPage(traversed.at(-1)).length) {
+        console.log(`Waypoint "${traversed.at(-1)}" has errors, so won't purge`);
+        return;
+      }
+
       const all = plan.getWaypoints();
       const toPurge = all.filter(e => !traversed.includes(e));
       journeyContext.purge(toPurge);
